chore: move react tests into dedicated folder
